calltr: special hack for REGISTER handling + timeouts
Hack to try catching related REGISTERs/REG replies and keeping
alive the REGISTER call entry for the duration of the registry
expiration (needed for reg-new/reg-del/reg-expire events
generation).
Only partially functional, REGISTER with changing to-tags are
still not handled properly. The contact mathing is for now
"optimistic" and not rfc conformant (parameters and headers in URI
are ignored).

The update state function returns now also the timeout (since the
timeout is not any longer a fixed function of state).