generate a reg-del for each binding on reg delete all
If a registration delete for all contacts (*) is received,
generate a reg-del for each removed bindings, even if reg-del are
configured to be generated immediately (reg_del_delay > 0). Note
that in the no reg-del delay, the reg-del for the 1st binding will
be generated immediately and the ones for the other matching
bindings after 1s.