fix adding contact to events
Commit 83b3c84a9d6f9f45bced7d2838d3d3f20dec01e6  changed
call-entries to store the whole contact and added a shortened
contact to the events. However the length check still used the
whole contact instead of the shortened one, causing a mismatch
for any contact with parameters added to an event and thus a
"truncated" event after the first such contact (resulting in
possible missing attributes like contact2, reason, ua, or pai).